+++ internal/types/types.go
@@ -3,7 +3,8 @@ package types
 import "github.com/pancsta/gosway/ipc"
 
 type WindowData struct {
-	ID        int
+	ID int
+	// eg HEADLESS-1
 	Output    string
 	Workspace string
 	Title     string