+++ CHANGELOG.md
@@ -1,13 +1,27 @@
+## v0.2.3
+
+- feat: add next-vnc use command
+- fix: fix titlebar-toggle for floating windows
+- fix: fix disabling mouse-follows-focused
+- fix: fix move focus to a divider (via workspace focus)
+- fix: always get outputs from IPC, not cache
+
+## v0.2.2
+
+- fix: add misc fixes
+- feat: add onCopy for user cmds
+- fix: fix clipboard in light mode
+
 ## v0.2.0
 
 - refac: rearrange under pressure
 - feat: add clipboard manager
 - feat: add support for user command files
 - fix: correct app IDs on startup
-- feat(usr-cmd): add titlebar-toggle
-- feat(usr-cmd): add resize-toggle
-- feat(usr-cmd): add arrange
-- feat(cmd): add win-to-space
+- feat: add titlebar-toggle
+- feat: add resize-toggle
+- feat: add arrange
+- feat: add win-to-space
 
 ## v0.1.0